Espers v0.8.5.3 Patch-6 supplementary 2
Updated dependencies, built new Windows client with updated Qt and
dependencies.

+++ src/db.cpp
@@ -47,7 +47,10 @@ void CDBEnv::EnvShutdown()
     if (ret != 0)
         LogPrintf("EnvShutdown exception: %s (%d)\n", DbEnv::strerror(ret), ret);
     if (!fMockDb)
-        DbEnv(0).remove(strPath.c_str(), 0);
+        // Removed for: Qt 5.8.0 | db-6.2.23.NC | boost 1.63 - build
+        // Reason: Causes error during build process due to overloaded function
+        // DbEnv(0).remove(strPath.c_str(), 0);
+        DbEnv((u_int32_t)0).remove(strPath.c_str(), 0);
 }
 
 CDBEnv::CDBEnv() : dbenv(DB_CXX_NO_EXCEPTIONS)
